JAMES-3459 fix package issue
