+++ server/data/data-jmap/src/main/java/org/apache/james/jmap/api/change/MailboxChangeRepository.java
@@ -23,8 +23,7 @@ import java.util.Optional;
 
 import org.apache.james.jmap.api.change.MailboxChange.Limit;
 import org.apache.james.jmap.api.change.MailboxChange.State;
-import org.apache.james.jmap.api.vacation.AccountId;
-import org.apache.james.jmap.memory.vacation.MemoryMailboxChangeRepository.MailboxChanges;
+import org.apache.james.jmap.api.model.AccountId;
 
 import reactor.core.publisher.Mono;
