+++ server/data/data-jmap/src/test/java/org/apache/james/jmap/api/change/MailboxChangeRepositoryContract.java
@@ -31,8 +31,7 @@ import java.util.UUID;
 import org.apache.james.jmap.api.change.MailboxChange.Limit;
 import org.apache.james.jmap.api.change.MailboxChange.State;
 import org.apache.james.jmap.api.exception.ChangeNotFoundException;
-import org.apache.james.jmap.api.vacation.AccountId;
-import org.apache.james.jmap.memory.vacation.MemoryMailboxChangeRepository.MailboxChanges;
+import org.apache.james.jmap.api.model.AccountId;
 import org.apache.james.mailbox.model.TestId;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
