+++ server/data/data-jmap/src/test/java/org/apache/james/jmap/api/vacation/NotificationRegistryContract.java
@@ -28,6 +28,7 @@ import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.james.jmap.api.model.AccountId;
 import org.apache.james.util.date.ZonedDateTimeProvider;
 import org.junit.jupiter.api.Test;
